// RR-Frontend/electron/main/enums/error-message.enum.ts
export enum ErrorMessage {
  APP_INITIALIZE_FAILED = 'Initializing app failed',
  PLATFORM_UNSUPPORTED = 'Platform not supported',
  PARSER_EXEC_NOT_EXIST = 'Dicom parser executable file does not exist at ',
  MISSING_PACS_CONFIG = 'Missing host or port information',
  READ_CONFIG_FAILED = 'Reading config file failed',
  CREATE_CONFIG_FAILED = 'Creating config.json failed',
  MISSING_BASEURL_CONFIG = 'Base url is not configured properly',
  FILE_LOG_FAILED = 'Log to file failed',
  LOAD_FAILED = 'Web contents failed to load',
  STUDY_SELECTION_CANCELLED = 'Study selection has been cancelled',
  MODULE_SELECTION_CANCELLED = 'Module selection has been cancelled',
  REQUIRED_FIELDS_CANCELLED = 'Prompting required fields has been cancelled',
  IMPORT_FAILED_TITLE = 'Fehler beim Import',
  IMPORT_FAILED_MESSAGE = 'Import fehlgeschlagen, siehe Logs in %APPDATA%\\RadioReport\\logs',
  NO_PARSER_DATA_MESSAGE = 'Die PACS Abfrage lieferte keine Daten',
  NO_PARSER_DATA = 'PACS returned no data',
  CMOVE_WITHOUT_STUDYID_MESSAGE = 'Für einen CMove Request wird die Angabe einer StudyInstanceUid benötigt',
  CMOVE_WITHOUT_STUDYID = 'CMove request sent without study instance uid',
  INVALID_PACS_CONFIG_MESSAGE = 'Ungültige PACS-Konfiguration, siehe Logs in %APPDATA%\\RadioReport\\logs',
  INVALID_PACS_CONFIG = 'Invalid PACS configuration',
  CERTIFICATE_ERROR = 'Certificate could not be verified',
  RENDERER_PROCESS_CRASHED = 'Renderer process unexpectedly disappeared/crashed',
  CHILD_PROCESS_CRASHED = 'Child process unexpectedly disappeared/crashed',
  PLUGIN_CRASHED = 'Plugin unexpectedly crashed',
  PRELOAD_ERROR = 'Error during preload script execution',
  INVALID_DICOM_MAP = 'Invalid dicom map',
  INVALID_DICOM_PARSER_RESULT = 'Invalid dicom parser result',
  INVALID_DICOM_QUERY_MODE = 'Invalid dicom query mode',
  MAINWINDOW_NOT_INITIALIZED = 'Main window is not initialized yet',
  DICOM_MAP_NOT_AVAILABLE = 'Dicom map has not been retrieved and set yet',
  SPECIAL_CHARACTERS_REMOVED = 'Special characters had been removed from cli argument',
  UNHANDLED_EXCEPTION = 'Uncaught exception occurred',
  UNHANDLED_REJECTION = 'Unhandled rejection occurred',
  RESULT_ARRAYS_DIFFERENT_LENGTH = 'Hex and keyword arrays of different length',
  NO_SERIES_DATA = 'No dicom parser data of series level retrieved',
  AUTO_UPDATE_FAILED = 'Auto update request failed',
  IMPORT_SCOPE_NOT_INITIALIZED = 'Import scope not initialized',
  MISSING_SCPPORT = 'CMove request without scp port'
}