// RR-Frontend/electron/main/enums/ipc-channel.enum.ts
export enum IpcChannel {
  UserLogin = 'user-login',
  ImportReportRequest = 'import-report-request',
  ToastError = 'toast-error',
  StudySelection = 'study-selection',
  ModuleSelection = 'module-selection',
  ReportRequiredFields = 'report-required-fields',
  RendererLog = 'renderer-log',
  ExistingImport = 'existing-import',
  GoToNextPage = 'go-to-next-page',
  GoToPreviousPage = 'go-to-previous-page',
  GoToReportPage = 'go-to-report-page'
}